# openapi_check.json
{"openapi": "3.1.0", "info": {"title": "OrchX Kernel Engine API", "description": "Universal interface and visualization layer for OrchX. Strict OpenAPI compliant schemas.", "version": "1.0.0"}, "paths": {"/api/v1/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "description": "Register a new workspace user.", "operationId": "register_api_v1_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "description": "Retrive JWT Access Token.", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["auth"], "summary": "Read User Me", "description": "Retrieve logged-in user.", "operationId": "read_user_me_api_v1_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/plugins/": {"get": {"tags": ["plugins"], "summary": "List Registered Plugins", "description": "Retrieve metadata of all active plugins loaded in the kernel registries.", "operationId": "list_registered_plugins_api_v1_plugins__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/DashboardPluginInfo"}, "type": "array", "title": "Response List Registered Plugins Api V1 Plugins  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/dashboard/stats": {"get": {"tags": ["dashboard"], "summary": "Get Minimized Telemetry", "description": "Retrieve minimized kernel metrics, health diagnostics, and logs.", "operationId": "get_minimized_telemetry_api_v1_dashboard_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DashboardStatsResponse"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/preview/render": {"post": {"tags": ["preview"], "summary": "Render Artifact for Preview Studio", "description": "Universal Visualization Layer Endpoint.\nTakes an artifact ID and type, returning the rendered standard JSON format\nfor the Preview Studio frontend.", "operationId": "render_artifact_api_v1_preview_render_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ArtifactPreviewRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ArtifactPreviewResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preview/history/{artifact_id}": {"get": {"tags": ["preview"], "summary": "Get Artifact History", "description": "Retrieve history of an artifact for diff/version previews.", "operationId": "get_artifact_history_api_v1_preview_history__artifact_id__get", "parameters": [{"name": "artifact_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Artifact Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ArtifactPreviewResponse"}, "title": "Response Get Artifact History Api V1 Preview History  Artifact Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/suggestions/": {"get": {"tags": ["suggestions"], "summary": "Get active suggestions", "description": "Retrieve ranked, advisory suggestions generated by OrchX subsystems.", "operationId": "get_suggestions_api_v1_suggestions__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuggestionResponse"}}}}}}}, "/api/v1/suggestions/{suggestion_id}/dismiss": {"post": {"tags": ["suggestions"], "summary": "Dismiss a suggestion", "description": "Dismiss a suggestion so it no longer appears in the active list.", "operationId": "dismiss_suggestion_api_v1_suggestions__suggestion_id__dismiss_post", "parameters": [{"name": "suggestion_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Suggestion Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/healthz": {"get": {"tags": ["health"], "summary": "Root Health Check", "description": "Verify kernel HTTP server availability.", "operationId": "root_health_check_healthz_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"ArtifactPreviewRequest": {"properties": {"artifact_id": {"type": "string", "title": "Artifact Id"}, "artifact_type": {"type": "string", "title": "Artifact Type"}}, "type": "object", "required": ["artifact_id", "artifact_type"], "title": "ArtifactPreviewRequest"}, "ArtifactPreviewResponse": {"properties": {"artifact_id": {"type": "string", "title": "Artifact Id"}, "artifact_type": {"type": "string", "title": "Artifact Type"}, "content": {"additionalProperties": true, "type": "object", "title": "Content"}, "metadata": {"additionalProperties": true, "type": "object", "title": "Metadata"}}, "type": "object", "required": ["artifact_id", "artifact_type", "content", "metadata"], "title": "ArtifactPreviewResponse"}, "AuditLogResponse": {"properties": {"action": {"type": "string", "title": "Action"}, "details": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Details"}, "id": {"type": "integer", "title": "Id"}, "user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "User Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["action", "id", "created_at"], "title": "AuditLogResponse"}, "Body_login_api_v1_auth_login_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "^password$"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "format": "password", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "format": "password", "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_api_v1_auth_login_post"}, "DashboardPluginInfo": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "version": {"type": "string", "title": "Version"}, "api_version": {"type": "string", "title": "Api Version"}, "type": {"type": "string", "title": "Type"}, "capabilities": {"items": {"type": "string"}, "type": "array", "title": "Capabilities"}, "permissions": {"items": {"type": "string"}, "type": "array", "title": "Permissions"}}, "type": "object", "required": ["id", "name", "version", "api_version", "type", "capabilities", "permissions"], "title": "DashboardPluginInfo"}, "DashboardStatsResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "version": {"type": "string", "title": "Version"}, "uptime": {"type": "number", "title": "Uptime"}, "loaded_components": {"items": {"type": "string"}, "type": "array", "title": "Loaded Components"}, "capabilities": {"items": {"type": "string"}, "type": "array", "title": "Capabilities"}, "providers_count": {"type": "integer", "title": "Providers Count"}, "agents_count": {"type": "integer", "title": "Agents Count"}, "tools_count": {"type": "integer", "title": "Tools Count"}, "workflows_count": {"type": "integer", "title": "Workflows Count"}, "recent_logs": {"items": {"$ref": "#/components/schemas/AuditLogResponse"}, "type": "array", "title": "Recent Logs"}}, "type": "object", "required": ["status", "version", "uptime", "loaded_components", "capabilities", "providers_count", "agents_count", "tools_count", "workflows_count", "recent_logs"], "title": "DashboardStatsResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Suggestion": {"properties": {"id": {"type": "string", "title": "Id", "description": "Unique identifier for the suggestion"}, "priority": {"$ref": "#/components/schemas/SuggestionPriority", "description": "Priority of the suggestion"}, "confidence": {"type": "number", "maximum": 1.0, "minimum": 0.0, "title": "Confidence", "description": "Confidence score between 0.0 and 1.0"}, "source": {"type": "string", "title": "Source", "description": "Source of the suggestion"}, "reasoning": {"type": "string", "title": "Reasoning", "description": "Reasoning behind the suggestion"}, "affected_components": {"items": {"type": "string"}, "type": "array", "title": "Affected Components", "description": "List of components affected"}, "estimated_impact": {"type": "string", "title": "Estimated Impact", "description": "Estimated impact of implementing this suggestion"}, "payload": {"anyOf": [{}, {"type": "null"}], "title": "Payload", "description": "Optional additional data for this suggestion"}}, "type": "object", "required": ["id", "priority", "confidence", "source", "reasoning", "estimated_impact"], "title": "Suggestion"}, "SuggestionPriority": {"type": "string", "enum": ["low", "medium", "high", "critical"], "title": "SuggestionPriority"}, "SuggestionResponse": {"properties": {"suggestions": {"items": {"$ref": "#/components/schemas/Suggestion"}, "type": "array", "title": "Suggestions"}}, "type": "object", "required": ["suggestions"], "title": "SuggestionResponse"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "UserCreate": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserCreate"}, "UserResponse": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "id": {"type": "integer", "title": "Id"}, "is_active": {"type": "boolean", "title": "Is Active"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["email", "id", "is_active", "created_at", "updated_at"], "title": "UserResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/v1/auth/login"}}}}}}