# packages/orchx-api/orchx_api/api/v1/suggestions.py
from fastapi import APIRouter, Depends, HTTPException
from typing import List
from pydantic import BaseModel
from orchx_core.interfaces.suggestion_contracts import Suggestion

router = APIRouter()

class SuggestionResponse(BaseModel):
    suggestions: List[Suggestion]

@router.get("/", response_model=SuggestionResponse, summary="Get active suggestions")
async def get_suggestions():
    """
    Retrieve ranked, advisory suggestions generated by OrchX subsystems.
    """
    # In a real environment, this would fetch from kernel.suggestion_engine.get_ranked_suggestions()
    # Mocking for now to satisfy OpenAPI schema generation.
    return SuggestionResponse(suggestions=[])

@router.post("/{suggestion_id}/dismiss", summary="Dismiss a suggestion")
async def dismiss_suggestion(suggestion_id: str):
    """
    Dismiss a suggestion so it no longer appears in the active list.
    """
    return {"status": "dismissed", "suggestion_id": suggestion_id}